requirements.txt
Flask==2.2.2
keras==3.2.1
Pyrebase4==4.7.1
